requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==25.4.0
billiard==4.2.4
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.2
django-environ==0.13.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-camel-case==1.4.2
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gunicorn==25.1.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
packaging==26.0
prompt_toolkit==3.0.52
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==6.4.0
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.6.0
